success to print edge of AVweight

+++ graph.hpp
@@ -19,7 +19,7 @@ struct vertex_property
 struct edge_property
 {
     std::string label;
-    double weight;
+    int weight;
 };
 
 // グラフの属性値を自由に定義
